Island: Move AwsCmdRunner to monkey_island/cc/server_utils/
