+++ monkey/monkey_island/cc/server_utils/aws_cmd_result.py
