+++ monkey/monkey_island/cc/server_utils/aws_cmd_runner.py
@@ -1,9 +1,9 @@
 import logging
 import time
 
-from common.cmd.aws.aws_cmd_result import AwsCmdResult
 from common.cmd.cmd_runner import CmdRunner
 from common.cmd.cmd_status import CmdStatus
+from monkey_island.cc.server_utils.aws_cmd_result import AwsCmdResult
 from monkey_island.cc.services import aws_service
 
 logger = logging.getLogger(__name__)
